Reorganise structure of the package
