--- agent/__init__.py
@@ -1,2 +1,0 @@
-from agent.rtm.client import AgentRTM
-from agent.web.client import AgentWeb
