+++ changelog.md
@@ -1,0 +1,16 @@
+# Change Log
+All notable changes to this project will be documented in this file.
+
+
+## [0.1.1] - 2021-04-06
+
+### Added
+
+- New header manipulation/retrieval methods [@DamianAmbrozkiewicz](https://github.com/DamianAmbrozkiewicz).
+- Changelog file [@kacperf531](https://github.com/kacperf531).
+- Simple usage example in `readme.md` [@kacperf531](https://github.com/kacperf531).
+
+### Changed
+
+- `agent`, `configuration` and `customer` packages now have common root `livechat`. [@kacperf531](https://github.com/kacperf531).
+- rename `ConfigurationApi` method for obtaining client to `get_client` for consistency with other clients [@kacperf531](https://github.com/kacperf531).
