--- configuration/__init__.py
@@ -1,1 +1,0 @@
-from configuration.client import ConfigurationApi
