--- customer/__init__.py
@@ -1,2 +1,0 @@
-from customer.web.client import CustomerWeb
-from customer.rtm.client import CustomerRTM
