+++ docs/agent_rtm.html
@@ -34,8 +34,8 @@
 import typing
 from abc import ABCMeta
 
-from utils.helpers import prepare_payload
-from utils.ws_client import WebsocketClient
+from livechat.utils.helpers import prepare_payload
+from livechat.utils.ws_client import WebsocketClient
 
 
 class AgentRTM:
