+++ docs/customer_rtm.html
@@ -33,8 +33,8 @@
 
 from abc import ABCMeta
 
-from utils.helpers import prepare_payload
-from utils.ws_client import WebsocketClient
+from livechat.utils.helpers import prepare_payload
+from livechat.utils.ws_client import WebsocketClient
 
 
 class CustomerRTM:
