+++ livechat/agent/__init__.py
@@ -1,0 +1,3 @@
+#pylint: disable=C0114
+from livechat.agent.rtm.client import AgentRTM
+from livechat.agent.web.client import AgentWeb
