+++ livechat/configuration/__init__.py
@@ -1,0 +1,2 @@
+#pylint: disable=C0114
+from livechat.configuration.client import ConfigurationApi
