+++ livechat/customer/__init__.py
@@ -1,0 +1,3 @@
+#pylint: disable=C0114
+from livechat.customer.rtm.client import CustomerRTM
+from livechat.customer.web.client import CustomerWeb
