+++ livechat/tests/test_configuration_api_client.py
@@ -4,41 +4,41 @@
 
 import pytest
 
-from configuration.client import ConfigurationApi
+from livechat.configuration.client import ConfigurationApi
 
 
 @pytest.fixture
 def conf_api_client():
     ''' Fixture returning Configuration API client. '''
-    return ConfigurationApi.get_api_client(token='test')
+    return ConfigurationApi.get_client(token='test')
 
 
-def test_get_api_client_without_args():
+def test_get_client_without_args():
     ''' Test if TypeError raised without args. '''
     with pytest.raises(TypeError) as exception:
-        ConfigurationApi.get_api_client()
+        ConfigurationApi.get_client()
     assert str(
         exception.value
-    ) == "get_api_client() missing 1 required positional argument: 'token'"
+    ) == "get_client() missing 1 required positional argument: 'token'"
 
 
-def test_get_api_client_without_token():
+def test_get_client_without_token():
     ''' Test if TypeError raised without token. '''
     with pytest.raises(TypeError) as exception:
-        ConfigurationApi.get_api_client(version='test')
+        ConfigurationApi.get_client(version='test')
     assert str(
         exception.value
-    ) == "get_api_client() missing 1 required positional argument: 'token'"
+    ) == "get_client() missing 1 required positional argument: 'token'"
 
 
-def test_get_api_client_with_non_existing_version():
+def test_get_client_with_non_existing_version():
     ''' Test if ValueError raised for non-existing version. '''
     with pytest.raises(ValueError) as exception:
-        ConfigurationApi.get_api_client(token='test', version='test')
+        ConfigurationApi.get_client(token='test', version='test')
     assert str(exception.value) == 'Provided version does not exist.'
 
 
-def test_get_api_client_with_valid_args(conf_api_client):
+def test_get_client_with_valid_args(conf_api_client):
     ''' Test if production API URL is used and token is added to headers for valid args. '''
     assert conf_api_client.api_url == 'https://api.livechatinc.com/v3.3/configuration/action'
     assert conf_api_client.session.headers.get('Authorization') == 'test'
