+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = lc-sdk-python
-version = 0.1.0
+version = 0.1.1
 description = Package which lets to work with LiveChat API.
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -20,7 +20,7 @@ classifiers =
     Programming Language :: Python :: 3.9
 
 [options]
-packages = find:
+packages=["livechat"]
 python_requires = >=3.6
 install_requires =
          websocket-client==0.57.0
@@ -29,5 +29,3 @@ install_requires =
 [options.packages.find]
 exclude =
          tests*
-
-
